Migrate tizen_device_discovery.dart to null safety (#295)
* Move TizenDeviceManager into a separate file

* Migrate tizen_device_discovery.dart to null safety
